commitao

+++ Crudizin_nada_basico/app/src/main/java/com/proway/crudizin_nada_basico/model/Doctor.kt
@@ -11,7 +11,7 @@ data class Doctor(
     @ColumnInfo(name = "name_doctor")
     val name_doctor: String,
 
-    val speciality_fk: Long,
+    val speciality_fk: Int,
 )
 data class DoctorWithSpeciality(
     @Embedded
